separated simulation from utils
Simulation was separated to make the code more readable
