+++ scripts/addons/cam/utils.py
@@ -866,77 +866,6 @@ def sampleChunksNAxis(o, pathSamples, layers):
 
     return chunks
 
-
-def createSimulationObject(name, operations, i):
-    oname = 'csim_' + name
-
-    o = operations[0]
-
-    if oname in bpy.data.objects:
-        ob = bpy.data.objects[oname]
-    else:
-        bpy.ops.mesh.primitive_plane_add(align='WORLD', enter_editmode=False, location=(0, 0, 0), rotation=(0, 0, 0))
-        ob = bpy.context.active_object
-        ob.name = oname
-
-        bpy.ops.object.modifier_add(type='SUBSURF')
-        ss = ob.modifiers[-1]
-        ss.subdivision_type = 'SIMPLE'
-        ss.levels = 6
-        ss.render_levels = 6
-        bpy.ops.object.modifier_add(type='SUBSURF')
-        ss = ob.modifiers[-1]
-        ss.subdivision_type = 'SIMPLE'
-        ss.levels = 4
-        ss.render_levels = 3
-        bpy.ops.object.modifier_add(type='DISPLACE')
-
-    ob.location = ((o.max.x + o.min.x) / 2, (o.max.y + o.min.y) / 2, o.min.z)
-    ob.scale.x = (o.max.x - o.min.x) / 2
-    ob.scale.y = (o.max.y - o.min.y) / 2
-    print(o.max.x, o.min.x)
-    print(o.max.y, o.min.y)
-    print('bounds')
-    disp = ob.modifiers[-1]
-    disp.direction = 'Z'
-    disp.texture_coords = 'LOCAL'
-    disp.mid_level = 0
-
-    if oname in bpy.data.textures:
-        t = bpy.data.textures[oname]
-
-        t.type = 'IMAGE'
-        disp.texture = t
-
-        t.image = i
-    else:
-        bpy.ops.texture.new()
-        for t in bpy.data.textures:
-            if t.name == 'Texture':
-                t.type = 'IMAGE'
-                t.name = oname
-                t = t.type_recast()
-                t.type = 'IMAGE'
-                t.image = i
-                disp.texture = t
-    ob.hide_render = True
-
-
-def doSimulation(name, operations):
-    """perform simulation of operations. Currently only for 3 axis"""
-    for o in operations:
-        getOperationSources(o)
-    limits = getBoundsMultiple(
-        operations)  # this is here because some background computed operations still didn't have bounds data
-    i = image_utils.generateSimulationImage(operations, limits)
-    cp = getCachePath(operations[0])[:-len(operations[0].name)] + name
-    iname = cp + '_sim.exr'
-
-    numpysave(i, iname)
-    i = bpy.data.images.load(iname)
-    createSimulationObject(name, operations, i)
-
-
 def extendChunks5axis(chunks, o):
     s = bpy.context.scene
     m = s.cam_machine
